Remove unnecessary tokenization, rename id rule (#338)
The parser and lexer no longer try to be specific about times and
abbreviations of times. They are replaced by a generic id, which can be
checked for validity. This reduces the complexity of the id rule and the
size of the generated code due to rule removal and DFA size reduction.


Closes: #337 
Progresses #275