TSQL: Implement full visit chain from highest grammar rule to SELECT (#363)
Leaving the default ANTLR generated implementations of the visit chain
in between tSqlFile and select statements mean that we had to override
the aggregateResult visitor, which happens because ANTLR will not be
able to know which alts were actually visited and so expects you to
decide wha to do with them.

We eliminate the problem by implement the visitors in the chain, though
they only deal with SELECT at the moment.

Closes: #356 
Progresses: #275