Implement XML functions such as query, and local id assignment (#373)
Implements two more elements of select lists, being 

`@localid = column`

and the various XML column related functions.

Both are additions to the  selectListElem visitor.

New IR is created that is used when the visitor recognizes XML functions
(and can expand to other types of functions). The IR indicates which
function is to be applied to which column expression.

progresses: #275 
progresses: #357