feat: Implement coverage tests for TSQL (#409)
Creates coverage tests for TSQL, in the same manner as the coverage for
Snowflake.

Coverage is a non-blocking test suite that allows us to measure how far
along the path of syntax coverage we are, and how much of the syntax we
are able to convert to DBSQL IR.

We capture parsing errors and check the IR to see if contains any
UnresolvedXXX, which flags constructs that we do not currently cover.

Closes: #406

Co-authored-by: Valentin Kasas <valentin.kasas@xebia.com>