feat: Starts the process of removing all the fluff from the lexers and parsers.
All the overdone keyword tokens and overspecfied oprtoin grammar rules are being
gradually removed to be replaced with tables that indicate valid options etc and
simple rules that look for XXX (= expression) and just pass all of them to the
parse tree walkers. Grammar is easier to maintain and will also result in faster
parsers.