Updates to developer documentation and add grammar formatting to maven (#490)
Here we update the documentation for developers to include grammar
formatting instructions and recognize dialects other than Snowflake.

We also implement grammar formatting in the Maven build cycle such that
the grammars are always formatted before ANTLR processes them, though
this does cause them to always be regenerated. This ensures the grammars
are easier to resolve in merge conflicts as well as being easier to
maintain.