TSQL: Implement remaining join pivot and apply operations (#506)
Here we implement:

 - CROSS JOIN (explicit)
 - CROSS APPLY
 - OUTER APPLY
 - PIVOT
 - UNPIVOT
 
The APPLY joins are peculiar to TSQL, but should be translatable to LEFT
LATERAL or LEFT LATERAL VIEW with an explode function. This however is
not going to be a simple translation in the manner of function to
function translation, and so we will defer this to Databricks SQL
generation, at least for now.