Implement TSQL FOR and OPTION clauses as part of SELECT and other statements (#496)
TSQL has clauses FOR and OPTION for SELECT, and later other statements
such as INSERT.

The FOR clause has no equivalent in Databricks SQL, but we must visit
the clause anyway, even though we build no IR for it. Later, it could be
a good idea to report this via the Error handler as unsupported (not
Unresolved), along with functions and other things that we are unable to
convert to anything sensible.

The OPTION clause provides TSQL with optimization hints for various
queries. We support parsing it and do build the options as they may be
useful to add as comments to the generated converted query as useful
hints in the migration process. In general, it is better to leave
optimization to the Catalyst optimizer. However, as options can be set
on the optimizer, knowing what hints were given to TSQL could be useful.