TSQL: Remove the SIGN fragment from numerical tokens (#547)
Fixes #546 

The proposed solution isn't completely satisfying though, as it makes it
impossible to parse negative literals out of TSQL input: the input
string `-2` is now parsed as `UMinus(Literal(2))` (instead of simply
`Literal(-2)`). While being a mere inconvenience (it just delegates the
handling of `UPlus` and `UMinus` to downstream processes), the fact that
Snowflake grammar is able to parse `-2` as `Literal(-2)` indicates that
there might be a better solution to this problem. So I suggest we keep
#546 open even if we merge this PR, so that @jimidle can later come back
to it and find a better solution.