Improve TSQL and Snowflake parser and lexer (#757)
Here, we correct a number of problems with the Snowflake lexer, which
was unnecessarily trying to make sense of things like escape sequences,
which are not meant to be verified in the lexer.

We also correct parsing of various options in Snowflake, which were all
spelled out as lexer tokens, when they are just simple strings. This
simplifies both lexer and parser rules.

Snowflake statements can accept `?` as placeholder for a future
expression bind. The Snowflake lexer and parser now accept this as a
`PARAM` placeholder in any expression. This seems to be most useful with
the `IDENTIFIER(?)` pseudo function, which protects against SQL
injection attacks.

The TSQL lexer has a few small improvements, mainly such that we use the
same name in Snowflake and TSQL for the lexer's catch-all rule, when an
input character matches no sequence specified in the lexer.

Fixes: #741