fix: Ensure that COLON id syntax is not duplicated
  - Not that the : is realy an operator here for JSON access:
      SELECT json_data:a

But this is not implemented yet, hence there are no tests, hence this
change was not picked up as being incorrect.

I have left the syntax spec as is for now until the JSON is implemented, where
a COLON expression should be `expression COLON expression`, as should
other constructs in Snowflake.