Rephrase implementation using a monadic `Result` (#892)
Incidentally, this helped fixing a bug where tests `transpile BANG with
negative scenario unknown command` and `transpile BANG with negative
scenario unknown command2` were mistakenly expected to return an empty
string.

The changes introduced by the original PR should have detected that as
wrong (these tests should actually raise an error), but failed to do so
because of a combination of `Option` fields in `Result` and
`.getOrElse("")` calls here and there.
# Conflicts:
#	core/src/main/scala/com/databricks/labs/remorph/transpilers/SnowflakeToDatabricksTranspiler.scala
#	core/src/test/scala/com/databricks/labs/remorph/transpilers/SnowflakeToDatabricksTranspilerTest.scala