Create more robust toolchain control with error recording for repeatable testing (#888)
Here, we extend the basic toolchain we had such that all errors are
captured and errors at any stage in the workflow
`parse->visit->optimize->transpile` prevent an attempt to continue the
workflow. This prevents, for example, attempts to generate IR from a
parse tree generated with syntax errors. There is no point trying to
generate IR from a broken ParseTree.

We also provide some basic source code providers for directories and
files.

Finally, we provide a command line interface for transpilation such that
developers can look at the results of any particular input. This
includes the ability to retrieve query history (Snowflake only for now),
attempt to compile each retrieved query and record the results as jsonl
output. The code the command line uses can be the basis for an ongoing
integration test further down the line.

---------

Co-authored-by: Valentin Kasas <valentin.kasas@xebia.com>