feat: Implement all of TSQL predicates except for SOME ALL ANY (#922)
Thee IR generation for TSQL predicates was not complete. Here, we
implement all of the IR for predicates such as IN, IS, BETWEEN, LIKE....

Additionally the LIKE IR was incorrectly assuming that the ESCAPE
character had to be a single character at parse time, but in fact it
accepts a full expression which must evaluate to a single character at
runtime. We correct that here.