Ensure that successful parsing is recorded even if transpilation fails and record reconciled generations (#920)
A small bug meant that if the transpile failed, then the fact that the
parse was successful was not recorded. Here we correct that oversight.

We also add an extra field to the report entry to record what percentage
of code generations resulted in the expected translation as currently we
record that there was a generation difference but do not report how many
resulted in the expected translation according to the coverage test.