Establish incomplete parser and aggregate functions for all Visitors (#966)
While we had implemented an IncompleteParser trait, it was somewhat
incompletely implemented. Here, we ensure that all visitors implement
the `ParserCommon` trait and supply the method `unresolved(String)`,
which ensure that null is never inserted in the IR.

We now provide logger warnings output to indicate that there are missing
visitors that are relying on default implementations and the text held
by an UnresolvedXXX node indicates which visitor was missing and what
source text the visitor is supposed to handle.

The main purpose of this change is to ensure that the IR is never
populated with null nodes or children so that IR visitors do not need to
check for these errors. All unimplemented visitors will be reported as
such and can then be implemented as we do not wish to rely on default
behavior.

The default implementation of every visitor is just a visit of all
children of the node. The visit then aggregates the results of all the
visits and returns the last thing visited, which happens to work in some
cases. This is the job of the `aggregateResult` method.

However, this is not the way to implement the visitors in a production
system as we should cover all elements of the parse tree with an actual
visitor that explicitly shows the intent of the visitor.

`wrapUnresovedInput` is still implemented and it can be used by concrete
implementations of visitors to show that we know we have not implemented
support for a particular ParseTree.

We also implement some of the required visitors that were missing to
date, although not all of them. we can gradually (or in one hit)
implement the reported missing visitors from here, but this is a big
enough change as it is.