Snowflake: Correct parsing of Array literals (#984)
Here we correct the ANTLR grammar and IR generation for Snowflake
arrayliterals of the form:

```sql
[2, 3, 4.00::DOUBLE, 4, NULL]
```
The existing grammar only accepted literal values for the array, but in
fact, because we must allow casts, the elements of the array need to be
parsed as expressions.

The current type determination in visitArrayLiteral may be too naive but
does not affect code generation as the generator does not use it. It
only determines the type of the array by looking at the first expression
in the array literal, but when an array literal contains a double (as
above) and the first value appears to be an integer, then the array
literal type should probably be typed as DoubleType and not IntegerType.
We will need a function that determines the highest base type that
covers all the values, if this ever becomes relevant. The code is
commented so that we do not forget this.


closes: #978