Create launcher from cli (#1315)
This PR enhances the CLI as follows:
- adds a `transpiler-config-path` arg to the CLI that points to a
pluggable transpiler configuration file
- defaults the transpiler-config-path` arg to 'sqlglot' such that
built-in transpilers can still be invoked during the transition phase
- makes the 'source-dialect' optional, it is only required if the
transpiler supports more than 1 dialect

It introduces an LSPEngine, which acts as an alternative to the existing
SqlglotEngine. It is not operational yet, this PR is about routing CLI
calls, not executing them.

The LSPEngine is configured from the configuration stored at the above
'transpiler' path.
The configuration file itself is a yaml file with 2 sections:
```
remorph:
  version: 1
  dialects:
    - snowflake
  environment:
    - SOME_ENV: abc
  command_line:
    - python
    - lsp_server.py
custom:
  whatever: xyz

```
The `remorph` section is versioned and must strictly follow our spec.
The `custom` section is free-style, for use by the LSP server itself.

Both engines are loaded by the CLI, and used throughout the code, for
both `transpile` and `lineage` methods.

Requires #1345 

Fixes #1298

---------

Co-authored-by: Guenia Izquierdo Delgado <guenia.izquierdo@databricks.com>