Fix pytest-cov crash (#1390)
The `pytest-cov` plugin is used to gather test coverage info, using the
`coverage` tool.
It suffers from a limitation where child processes created by tests do
not inherit the `coverage` configuration from `pyproject.toml` because
that file is not on the command line.
This limitation is described here
https://github.com/nedbat/coveragepy/issues/512
With the introduction of LSP, we started encountering crashes when
collecting coverage data, because LSP launches a child process.
The solution is to move `coverage` configuration into a dedicated
`.coveragerc` file.

This PR does just that.

Fixes #1363

Co-authored-by: Guenia Izquierdo Delgado <guenia.izquierdo@databricks.com>