Configure transpile errors file (#1408)
Our current implementation hard codes the location of the file used to
write transpile errors.
This is not in line with other transpile flags that the user can
override.
It also exhibits undesirable side effects where error files are created
in the source code hierarchy when running unit tests.

This PR fixes the issue by adding an optional error-file flag that users
can provide.
It also renames the default error file name to use a timestamp rather
than a pid.
It also updates tests that generate error files to put them in a tmp
folder

Requires #1410

---------

Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>