Introduce table definition (#1450)
This pull request introduces significant enhancements to the table
discovery module by adding new data structures and services for geting
table definitions. The changes include the creation of data classes for
table and field information, the introduction of an abstract service for
fetching table definitions, and a concrete implementation for T-SQL
table definitions.

---------

Co-authored-by: sundar.shankar <sundar.shankar@databricks.com>
Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>