Freshen project YARN dependencies (used to build documentation) (#1525)
This PR updates all of the (frozen) dependencies that Yarn uses when
building project documentation. Aside from being a routine chore, this
should also resolve the security alerts that Dependabot has logged
against this project. (It also supersedes a bunch of PRs that Dependabot
has open, which I'll deal with after this.)