Update transpiler configurations to use a dictionary to specify the environment (#1529)
This PR updates the way environment variables are specified in our
transpiler configurations. Specifically the configuration now takes a
mapping of environment variables, rather than a list of mappings that
are merged during load.

Although an incompatible change:

- Our transpiler configurations don't use this part of the configuration
file;
 - We haven't yet released.

As such I think it's worth the benefit of tidying this up before we make
a release.