Fix test temporary resources (#1548)
This PR updates our tests to use the [pytest
facility](https://docs.pytest.org/en/6.2.x/tmpdir.html#the-tmp-path-fixture)
for creating temporary directories which tests can use. The primary
purpose of this change is to resolve #1547: something in the way the
existing code worked was (on macOS) creating a spurious `n` folder in
the root of the project.

Some incidental changes in this PR include:

 - Type annotations for the updated tests.
 - Avoiding unnecessary conversion of paths to strings.
- Failing if for some reason one of the temporary paths being created
already exists.