Mark `test_execute_query_without_connection` as allowed to fail locally if ODBC isn't available (#1550)
This PR resolves #1549 by marking the
`test_execute_query_without_connection` test as expected to fail, but
only locally and when ODBC isn't available.

There's currently an undocumented platform dependency that requires the
system to have native ODBC libraries installed for this test to pass.
This cannot be satisfied by Hatch, and rather than complicate setup
instructions, this PR marks the test as expected to fail, but only:

 - If running locally; and
 - The `pyodbc` module cannot be imported.

Note that under CI this test must pass, and currently does because the
CI environment sets up ODBC before running the tests.