Constrain the Databricks SDK to a single version (#1551)
As part of #1523 most dependencies were constrained to the specific
version currently in use. Unfortunately I overlooked the version for
`databricks-sdk` and left it as a wide range rather than pinning it to
the latest release (or a patch-release thereof).

This PR fixes this by pinning the `databricks-sdk` to a single version,
plus any patch releases that are made for it.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>