Add Pipeline and Step Comment Attributes (#1557)
## Changes

This PR adds a `comment` attribute to the `PipelineConfig` and `Step`
classes.

### What does this PR do? 

Adds the ability for pipeline authors to document what a pipeline and/or
step does for future maintainers (including themselves).

### Relevant implementation details

For pipelines containing many steps, it becomes unclear what effect a
pipeline or step has. To discern what a step does, for example, a
pipeline maintainer must look at the extract_source attribute, navigate
to the Python or SQL file, and then read through the file to determine
what a step does.

### Caveats/things to watch out for when reviewing:

### Linked issues

Resolves #1468. 

### Functionality

- [ ] added relevant user documentation
- [ ] added new CLI command
- [ ] modified existing command: `databricks labs remorph ...`
- [X] Augments the `PipelineConfig` and `Step` classes with a metadata
attribute.

### Tests

- [ ] manually tested
- [x] added unit tests
- [X] added integration tests

Co-authored-by: SundarShankar89 <72757199+sundarshankar89@users.noreply.github.com>