Wait for transpile capability during LSP initialisation (#1571)
## Changes

This PR updates the initialisation sequence when starting up the LSP
server: initialisation now waits until it knows the transpiler
capability is available before returning.

### Relevant implementation details

Currently we are using LSPs dynamic registration mechanism for the
server (=transpiler plugin), after the initialisation sequence, to
notify the client (=remorph CLI) that transpilation is supported. This
normally happens quickly, but previously the first call to transpile
would wait up to 10 seconds for the capability to be declared. As noted
on #1542, this meant that a failure that prevented it from arriving
would be noted on the call to transpile, which is a little misleading
because something has really gone wrong during startup.

### Caveats/things to watch out for when reviewing:

We don't currently have any way to verify the sequence of LSP
interactions at a fine-grained level: this would require mocking the LSP
part of the stack, and be a significant amount of work. Instead the
approach until now is to use a small test LSP server. Testing for this
PR is therefore a little racy: we can't strictly verify that
initialisation waits for the capability, because it might occur after
initialisation but before the test performs its check. A passing test
therefore doesn't _prove_ that it works in the strict sense, although a
failure (including flakiness) would demonstrate that it's not working as
intended.

### Linked issues

Resolves #1542.

### Tests

- update tests