LSP: include the client's process ID during the initialisation sequence (#1574)
## Changes

This PR updates the LSP initialisation sequence for pluggable
transpilers so that the client (=Remorph CLI) notifies the server
(=pluggable transpiler) of its process ID. The [LSP
specification](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#initialize)
mentions this:
```typescript
/**
 * The process Id of the parent process that started the server. Is null if
 * the process has not been started by another process. If the parent
 * process is not alive then the server should exit (see exit notification)
 * its process.
 */
```

The intent here is that sometimes a server cannot rely on orderly
shutdown of the LSP server (including the LSP messages or stdin/stdout
being closed) to detect a client that has crashed, and a mitigation here
is that the server should periodically check that this process is still
present.

### Caveats/things to watch out for when reviewing:

There's a line from another test removed: this is from a test added in
#1571, and #1569 would have removed the line but both of those PRs were
merged in parallel.

### Functionality

 - updated initialisation sequence during transpile requests.

### Tests

 - added tests