Improve logging during transpiler installation (#1576)
## Changes

This PR updates the logging while installing the transpilers. The goals
of these changes are:

 - Less logging during normal routine installation.
 - Correct some misleading logging.
- Improve the information available when things go wrong, to help users
understand the cause of a problem and our ability to help them.

### Relevant implementation details

Technical changes include:

 - Removing a confusing message when installing the components:
   ```
   INFO: Installing Remorph v0.9.0
   ```
   Remorph is already installed at this point, so it made no sense.
- Downgrading some logging from `INFO` to `DEBUG`. The goal here is
reduce the normal logging so that there is less noise during routine
progress. If we log before and after something that might fail, I have
chosen to log at `DEBUG` before and `INFO` upon completion although
either would be fine. (If you object, please say so.)
- Ensuring that failures to contact PyPi or Maven Central are logged.
These are early in the process, and will be a common source of failure.
To support users (and let them help themselves) it is important that the
reason for the failure be available.
- Ensuring that an `ERROR` is logged on all the paths whereby we wanted
to install a transpiler but this didn't happen.
- Improved logging during the initial install if the component metadata
can't be downloaded. This is really a failure, but previously we would
log:
   ```
INFO: databricks-labs-remorph-community-transpiler vNone is already
installed.
   ```
   or
   ```
   INFO: Databricks Morpheus transpiler vNone already installed
   ```
   Now we instead log:
   ```
WARN: Could not determine the latest version of:
databricks-labs-remorph-community-transpiler
   ```
   and
   ```
WARN: Could not determine the latest version of Databricks Morpheus
transpiler (com.databricks.labs:morpheus)
   ```
respectively. Both warnings are now also followed by an `ERROR`
indicating that installation failed.
- Some type hints have been added in a few places; this was purely
opportunistic.

### Caveats/things to watch out for when reviewing:

The PyPi and Maven installation code isn't symmetrical, so the logging
needs to be changed in multiple places.

To run the installer:

1. Wipe your remorph installation:
   ```sh
   rm -fr ~/.databricks/labs/remorph*
   ```
2. Install this branch:
   ```sh
   databricks labs install remorph@improve-install-logging
   ```
3. Work around an installer issue:
   ```sh
mkdir -p
~/.databricks/labs/remorph-transpilers/remorph-community-transpiler
   ```
4. Install the components:
   ```sh
   databricks labs remorph install-transpile
   ```

### Functionality

- modified existing command: `databricks labs remorph install-transpile`

### Tests

- manually tested

Some example logs from a clean install:
```
17:30:04  INFO [d.l.remorph.install] Successfully installed databricks-labs-remorph-community-transpiler v0.0.1
17:30:04  WARN [d.l.remorph.install] Could not determine the latest version of Databricks Morpheus transpiler (com.databricks.labs:morpheus)
17:30:04 ERROR [d.l.remorph.install] Failed to install transpiler: Databricks Morpheus transpiler
17:30:04  INFO [d.l.remorph.install] Configuring remorph `transpile`.
```

And a second run with the components already installed:
```
17:31:59  INFO [d.l.remorph.install] databricks-labs-remorph-community-transpiler v0.0.1 already installed
17:31:59  WARN [d.l.remorph.install] Could not determine the latest version of Databricks Morpheus transpiler (com.databricks.labs:morpheus)
17:31:59 ERROR [d.l.remorph.install] Failed to install transpiler: Databricks Morpheus transpiler
```