Continued: flexible transpiler configuration (#1594)
This PR continues from #1488, introducing the following:

- The `--debug` handling introduced in #1488 is removed. This was a
workaround for a problem solved upstream in databrickslabs/remorph#238.
- The logging has been adjusted to reflect the updates we made in #1576.
- A test from #1568 is fixed; something on this branch leads the version
information to include a timestamp, which needs to be ignored by that
test.
- Some existing (installation) tests were incorrectly mocking the
install-state: the values should always be strings, but the tests were
placing integers in the fixture. Some upstream changes on
databrickslabs/remorph#189 change the behaviour in this area.

### Implementation details

~This pull-request depends on databrickslabs/blueprint#241, and the
project dependencies have been updated to depend on them. This is a
temporary change to ensure that CI can work. Before merging a release of
blueprint is required, and the released version needs to be used
instead.~ [Done]

### Caveats/things to watch out for when reviewing:

Please check the changes to the `transpile` command.

### Linked issues

Follows #1488.

### Functionality

- modified existing command: `databricks labs remorph transpile` (from
#1488)

### Tests

- existing unit tests
- added integration tests

---------

Co-authored-by: Eric Vergnaud <eric.vergnaud@databricks.com>
Co-authored-by: sundar.shankar <sundar.shankar@databricks.com>