Update maven installer to not use `search.maven.org` (#1611)
## Changes

### What does this PR do?

This PR updates `MavenInstaller` so that it no longer uses
`search.maven.org` to either:

 - Locate the current version of an artefact; or
 - Download the actual artefact.

### Relevant implementation details

The installer now fetches the metadata needed for detecting the version,
as well as the artefact itself, from the maven central repository
directly. This is the CDN-backed service where artefacts are supposed to
be fetched from, and it is much more reliable than the search service.

In addition to this, the installer had some logic that could abort the
installation if a file already exists at the target location on the
local filesystem. This check now happens before the download so that it
aborts immediately rather than first performing the (ultimately futile)
download.

### Caveats/things to watch out for when reviewing:

In addition to using the repository directly, this PR starts plumbing
through support for providing the artefact classifier. ~Another PR will
be needed to:~

~1. Expose the classifier as a parameter.~
~2. Update `install_morpheus()` to provide the classifier.~

~(Without these the normal installer will fail even after we publish
morpheus because we're not currently specifying the classifier.)~

[It looks like we're not publishing with a classifier, so this isn't
needed.]

### Linked issues

Resolves #1602 

### Functionality

- modified existing command: `databricks labs remorph install-transpile`

### Tests

- updated integration tests