Fix timestamp in morpheus transpiler state file (#1615)
## Changes

### What does this PR do? 

This PR fixes a bug in the morpheus installer whereby the state-file
that it writes has a timestamp that isn't in the expected format: it
should include timezone information.

### Relevant implementation details

The same issue was caught on the PyPi side during review of the initial
implementation but was missed on the Maven side. This PR factors out the
code so that it's shared between the implementations, and also adds a
unit-test to verify it.

### Caveats/things to watch out for when reviewing:

To deal with some type-checking issues, this PR also removes the
`_current_version` and `_latest_version` properties from the installers.
It was difficult to reason about whether they had been set properly (or
were maybe `None`) so plumbing them through the call-chain lets the
type-checker eliminate state-related bugs in this area.

### Functionality

- modified existing command: `databricks labs remorph install-transpile`

### Tests

- manually tested
- added unit tests
- existing integration tests