Properly pin the Databricks SDK. (#1639)
## Changes
### What does this PR do? 

This PR properly pins the project to a release of the Databricks SDK (or
patch releases), which happens to resolve a linting error on `main` due
to a breaking change in a later SDK release.

### Caveats/things to watch out for when reviewing:

The `~=` syntax means the _last_ digit can be incremented safely.
Because the `.0` was missing, we were allowing `0.51+` instead of
`0.51.x`.