Improve CLI 'transpile' UX (#1582)
In order to transpile code, the user must call:
`databricks labs remorph transpile ...`
Our current implementation requires the user to pass all `transpile`
config options as parameters to the above command, which is painful, and
leads to wonder what's the point in configuring the transpiler...
PR #1488 makes it less painful by making all parameters optional, and
using their configured counterpart if not provided.
However, if the resulting config is inconsistent, it will fail, which
can be frustrating.
This PR goes a step further by prompting the user if a config value is
missing or inconsistent.
Unit tests are provided, but this hasn't been tested manually.

Fixes #1538 
Requires #1488

---------

Co-authored-by: Andrew Snare <asnare@users.noreply.github.com>
Co-authored-by: Guenia Izquierdo Delgado <guenia.izquierdo@databricks.com>
Co-authored-by: sundar.shankar <sundar.shankar@databricks.com>
Co-authored-by: Andrew Snare <andrew.snare@databricks.com>