Make transpiler configuration path optional in the workspace transpile configuration (#1641)
## Changes

### What does this PR do? 

This PR makes the `transpiler_path_config` setting optional in
`TranspilerConfig`, normally set during the post-install configuration
questionnaire as part of the `install-transpile` subcommand.

This is in turn means we don't need to store `"None"` (the string) as a
value in the file, in turn meaning we don't need downstream workarounds
when reading the configuration.

### Relevant implementation details

In addition, when upgrading a v2 configuration we no longer need to set
a value that will fail: we can just leave it out. Even though this is
uncommon, leaving it unfilled rules out a source of issues downstream.

### Caveats/things to watch out for when reviewing:

Making an existing field optional is forward-compatible, so we don't
need to introduce a new version of the configuration.

### Linked issues

Resolves #1640.

Relates to #1637.

### Functionality

- modified existing command: `databricks labs remorph install-transpile`

### Tests

- manually tested
- existing tests