[BUGFIX] add test and change behavior for config option (#1662)
## Changes
Our current code always returns a value when prompting the user with a
question.
We need a mechanism where it can return None instead.

### What does this PR do? 
This PR implements a mechanism that when the default value for a prompts
question is `<none>`, if the user enters `<none>` or nothing, then the
return value is None

### Relevant implementation details
This is only available for LSPConfigOption class

### Linked issues
Requires https://github.com/databrickslabs/bladerunner/pull/289

### Tests
- [x] not tested

---------

Co-authored-by: Guenia Izquierdo Delgado <guenia.izquierdo@databricks.com>