Documentation markdown improvements (#1688)
## Changes

### What does this PR do?

This PR contains addresses some cosmetic issues in the documentation.
These include:

 - A typo in the _Overview_ section.
- Formatting issues for lists in the _Assessment_ section, including
line-breaks.
 - A few stray `remorph` references in example commands.
 - A broken image link. 

### Caveats/things to watch out for when reviewing:

Markdown is sensitive to trailing whitespace: two trailing space
characters at the end of a line insert a line-break. These are very
difficult to spot, and our repository settings try to strip trailing
whitespace. As such it's better to use an explicit `<br/>` tag where a
line-break is intended.

### Functionality

- updated user documentation

### Tests

- manually reviewed